# lab/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Turwin Medical Center API",
    "description": "API documentation for Turwin Medical Center's hospital management system",
    "version": "1.0.0",
    "contact": {
      "name": "Security Team",
      "email": "security@turwinmedical.example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:1730",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/login"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "User's username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "User's password"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password", "role"],
        "properties": {
          "username": {
            "type": "string",
            "description": "Desired username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "User's password"
          },
          "role": {
            "type": "string",
            "enum": ["doctor", "patient"],
            "description": "User's role"
          }
        }
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique user identifier"
          },
          "username": {
            "type": "string",
            "description": "User's username"
          },
          "role": {
            "type": "string",
            "enum": ["doctor", "patient"],
            "description": "User's role in the system"
          },
          "profile_image": {
            "type": "string",
            "description": "URL to user's profile image"
          }
        }
      },
      "TestResultRequest": {
        "type": "object",
        "required": ["patient_id", "doctor_id", "test_type", "result"],
        "properties": {
          "patient_id": {
            "type": "integer"
          },
          "doctor_id": {
            "type": "integer"
          },
          "test_type": {
            "type": "string"
          },
          "result": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "TestResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique test result identifier"
          },
          "patient_id": {
            "type": "integer",
            "description": "ID of the patient"
          },
          "doctor_id": {
            "type": "integer",
            "description": "ID of the doctor who ordered the test"
          },
          "test_type": {
            "type": "string",
            "description": "Type of medical test"
          },
          "result": {
            "type": "string",
            "description": "Test result details"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "Date of the test"
          },
          "image_path": {
            "type": "string",
            "description": "Path to any associated test images"
          }
        }
      }
    }
  },
  "paths": {
    "/api/login": {
      "post": {
        "summary": "Authenticate user",
        "description": "Authenticate user and return JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT token for authentication"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "summary": "Register new user",
        "description": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g., username taken)"
          }
        }
      }
    },
    "/api/users/patients": {
      "get": {
        "summary": "Get all patients",
        "description": "Retrieve a list of all patients. Only accessible by doctors.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of patients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized - doctors only"
          }
        }
      }
    },
    "/api/users/{userId}": {
      "get": {
        "summary": "Get user details",
        "description": "Retrieve user information. Users can only access their own information unless they are doctors.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized to access this user's data"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "patch": {
        "summary": "Update user profile",
        "description": "Update user information. Users can only update their own information.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/test-results": {
      "post": {
        "summary": "Create test result",
        "description": "Create a new test result. Only accessible by doctors.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestResultRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Test result created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResult"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized - doctors only"
          }
        }
      }
    },
    "/api/test-results/{patientId}": {
      "get": {
        "summary": "Get patient test results",
        "description": "Get all test results for a patient. Only accessible by doctors or the patient themselves.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of test results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestResult"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized to access these test results"
          }
        }
      }
    },
    "/api/upload-image/{userId}": {
      "post": {
        "summary": "Upload image",
        "description": "Upload an image for a user or test result. Users can only upload images for themselves.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "test_result_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "image_path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Authentication",
      "description": "User authentication endpoints"
    },
    {
      "name": "Users",
      "description": "User management endpoints"
    },
    {
      "name": "Test Results",
      "description": "Medical test results management"
    },
    {
      "name": "Images",
      "description": "Image upload endpoints"
    }
  ],
  "security": []
}